+++ src/renderer/src/ui/views/Main/{fragments}/Sidebar/model/types.ts
@@ -1,1 +1,0 @@
-export type SidebarSection = "calendar" | "tags" | "themes" | "deleted"